Move api/ to proto/
- All test variations pass.
