+++ internal/hermes-api/session/web_token_test.go
@@ -10,9 +10,9 @@ import (
 	"time"
 
 	"github.com/google/uuid"
-	"github.com/ownmfa/hermes/api/go/token"
 	"github.com/ownmfa/hermes/pkg/crypto"
 	"github.com/ownmfa/hermes/pkg/test/random"
+	"github.com/ownmfa/hermes/proto/go/token"
 	"github.com/ownmfa/proto/go/api"
 	"github.com/stretchr/testify/require"
 	"google.golang.org/protobuf/proto"
