+++ internal/hermes-api/test/identity_test.go
@@ -12,11 +12,11 @@ import (
 	"time"
 
 	"github.com/google/uuid"
-	"github.com/ownmfa/hermes/api/go/message"
 	ikey "github.com/ownmfa/hermes/internal/hermes-api/key"
 	"github.com/ownmfa/hermes/pkg/key"
 	"github.com/ownmfa/hermes/pkg/oath"
 	"github.com/ownmfa/hermes/pkg/test/random"
+	"github.com/ownmfa/hermes/proto/go/message"
 	"github.com/ownmfa/proto/go/api"
 	"github.com/stretchr/testify/require"
 	"google.golang.org/protobuf/proto"
