+++ proto/go/token/doc.go
